zigbee/requirements.txt
pandas
numpy
plotly
kaleido
pyserial
xbee
